refactor: 最後の媒体は外せない仕様に簡素化し media-selection を model 配下へ移動
- media-multi-filter: 空選択を作らないよう、選択が1件のときはその媒体を
  外せない（no-op）ようにしてトグルを簡素化。全メディアへ丸める分岐を撤去
- FSD に倣い media-selection を features/article/model/ 配下へ移動
